+++ api/auth/__init__.py
@@ -1,0 +1,17 @@
+# Authentication module for text2sql API
+
+from .user_management import (
+    ensure_user_in_organizations,
+    update_identity_last_login,
+    validate_and_cache_user,
+    token_required
+)
+from .oauth_handlers import setup_oauth_handlers
+
+__all__ = [
+    "ensure_user_in_organizations",
+    "update_identity_last_login", 
+    "validate_and_cache_user",
+    "token_required",
+    "setup_oauth_handlers"
+]
